WIP exploring building uiSelect and uiMention sub-components
I would like to explore this design pattern for the future of
ui-select, but the major feature I need is multiple select and
ui-select is terribly bloated. I am hoping that laying this groundwork
will start a pattern of highly modular and versatile select
components that can keep the code light at the same time.